Sliver app bar for album page
